requirements.txt
cs50
Flask
Flask-Session
requests
urllib.request
json
csv
werkzeug.security